@@ -47,7 +47,7 @@ func TestCloudFoundryCreateSpace(t *testing.T) {
 			s.ShouldFailOnCommand = nil
 		}()
 
-		s.ShouldFailOnCommand = map[string]error{"yes '' | cf login -a https://api.endpoint.com -u testUser -p testPassword ": fmt.Errorf(errorMessage)}
+		s.ShouldFailOnCommand = map[string]error{"yes '' | cf login -a https://api.endpoint.com -u testUser -p testPassword ": fmt.Errorf("%s", errorMessage)}
 
 		e := runCloudFoundryCreateSpace(&config, &telemetryData, cf, &s)
 		assert.EqualError(t, e, "Error while logging in occured: "+errorMessage)
@@ -67,7 +67,7 @@ func TestCloudFoundryCreateSpace(t *testing.T) {
 		defer cfUtilsMock.Cleanup()
 		errorMessage := "cf space creation error"
 
-		m.ShouldFailOnCommand = map[string]error{"cf create-space testSpace -o testOrg": fmt.Errorf(errorMessage)}
+		m.ShouldFailOnCommand = map[string]error{"cf create-space testSpace -o testOrg": fmt.Errorf("%s", errorMessage)}
 
 		cfUtilsMock.LoginError = errors.New(errorMessage)
 
