@@ -44,10 +44,8 @@ func (p *contrastExecuteScanReports) persist(stepConfig contrastExecuteScanOptio
 		{FilePattern: "**/toolrun_contrast_*.json", ParamRef: "", StepResultType: "contrast"},
 		{FilePattern: "**/piper_contrast_report.json", ParamRef: "", StepResultType: "contrast"},
 	}
-	envVars := []gcs.EnvVar{
-		{Name: "GOOGLE_APPLICATION_CREDENTIALS", Value: gcpJsonKeyFilePath, Modified: false},
-	}
-	gcsClient, err := gcs.NewClient(gcs.WithEnvVars(envVars))
+
+	gcsClient, err := gcs.NewClient(gcpJsonKeyFilePath, "")
 	if err != nil {
 		log.Entry().Errorf("creation of GCS client failed: %v", err)
 		return
