@@ -172,10 +172,8 @@ func (p *fortifyExecuteScanReports) persist(stepConfig fortifyExecuteScanOptions
 		{FilePattern: "**/piper_fortify_report.json", ParamRef: "", StepResultType: "fortify"},
 		{FilePattern: "**/piper_fortify_report.html", ParamRef: "", StepResultType: "fortify"},
 	}
-	envVars := []gcs.EnvVar{
-		{Name: "GOOGLE_APPLICATION_CREDENTIALS", Value: gcpJsonKeyFilePath, Modified: false},
-	}
-	gcsClient, err := gcs.NewClient(gcs.WithEnvVars(envVars))
+
+	gcsClient, err := gcs.NewClient(gcpJsonKeyFilePath, "")
 	if err != nil {
 		log.Entry().Errorf("creation of GCS client failed: %v", err)
 		return
