@@ -59,7 +59,7 @@ func runIntegrationArtifactDownload(config *integrationArtifactDownloadOptions, 
 	contentDisposition := downloadResp.Header.Get("Content-Disposition")
 	disposition, params, err := mime.ParseMediaType(contentDisposition)
 	if err != nil {
-		return errors.Wrapf(err, "failed to read filename from http response headers, Content-Disposition "+disposition)
+		return errors.Wrapf(err, "failed to read filename from http response headers, Content-Disposition %s", disposition)
 	}
 	filename := params["filename"]
 
@@ -71,12 +71,12 @@ func runIntegrationArtifactDownload(config *integrationArtifactDownloadOptions, 
 		workspaceRelativePath := config.DownloadPath
 		err = os.MkdirAll(workspaceRelativePath, 0755)
 		if err != nil {
-			return errors.Wrapf(err, "Failed to create workspace directory")
+			return errors.Wrap(err, "Failed to create workspace directory")
 		}
 		zipFileName := filepath.Join(workspaceRelativePath, filename)
 		file, err := os.Create(zipFileName)
 		if err != nil {
-			return errors.Wrapf(err, "Failed to create integration flow artifact file")
+			return errors.Wrap(err, "Failed to create integration flow artifact file")
 		}
 		if _, err := io.Copy(file, downloadResp.Body); err != nil {
 			return err
