@@ -48,10 +48,8 @@ func (p *malwareExecuteScanReports) persist(stepConfig malwareExecuteScanOptions
 		{FilePattern: "**/toolrun_malwarescan_*.json", ParamRef: "", StepResultType: "malwarescan"},
 		{FilePattern: "", ParamRef: "reportFileName", StepResultType: "malwarescan"},
 	}
-	envVars := []gcs.EnvVar{
-		{Name: "GOOGLE_APPLICATION_CREDENTIALS", Value: gcpJsonKeyFilePath, Modified: false},
-	}
-	gcsClient, err := gcs.NewClient(gcs.WithEnvVars(envVars))
+
+	gcsClient, err := gcs.NewClient(gcpJsonKeyFilePath, "")
 	if err != nil {
 		log.Entry().Errorf("creation of GCS client failed: %v", err)
 		return
@@ -252,11 +250,6 @@ func malwareExecuteScanMetadata() config.StepData {
 							{
 								Name:  "commonPipelineEnvironment",
 								Param: "custom/dockerConfigJSON",
-							},
-
-							{
-								Name: "dockerConfigJsonCredentialsId",
-								Type: "secret",
 							},
 
 							{
