@@ -471,7 +471,7 @@ func handleHTTPError(resp *http.Response, err error, message string, connectionD
 		match, _ := regexp.MatchString(".*EOF$", err.Error())
 		if match {
 			AddDefaultDashedLine(1)
-			log.Entry().Infof("%s", "A connection could not be established to the ABAP system. The typical root cause is the network configuration (firewall, IP allowlist, etc.)")
+			log.Entry().Info("A connection could not be established to the ABAP system. The typical root cause is the network configuration (firewall, IP allowlist, etc.)")
 			AddDefaultDashedLine(1)
 		}
 
