@@ -30,7 +30,7 @@ func Decrypt(secret, base64CipherText []byte) ([]byte, error) {
 	iv := cipherText[:aes.BlockSize]
 	cipherText = cipherText[aes.BlockSize:]
 
-	stream := cipher.NewCFBDecrypter(block, iv)
+	stream := cipher.NewCTR(block, iv)
 	stream.XORKeyStream(cipherText, cipherText)
 
 	return cipherText, nil
@@ -54,7 +54,7 @@ func Encrypt(secret, inBytes []byte) ([]byte, error) {
 		return nil, fmt.Errorf("failed to init iv: %w", err)
 	}
 
-	stream := cipher.NewCFBEncrypter(block, iv)
+	stream := cipher.NewCTR(block, iv)
 	stream.XORKeyStream(cipherText[aes.BlockSize:], inBytes)
 
 	return []byte(base64.StdEncoding.EncodeToString(cipherText)), nil
