@@ -2,7 +2,10 @@
 
 package mocks
 
-import mock "github.com/stretchr/testify/mock"
+import (
+	"context"
+	mock "github.com/stretchr/testify/mock"
+)
 
 // Client is an autogenerated mock type for the Client type
 type Client struct {
@@ -24,7 +27,7 @@ func (_m *Client) Close() error {
 }
 
 // DownloadFile provides a mock function with given fields: bucketID, sourcePath, targetPath
-func (_m *Client) DownloadFile(bucketID string, sourcePath string, targetPath string) error {
+func (_m *Client) DownloadFile(ctx context.Context, bucketID string, sourcePath string, targetPath string) error {
 	ret := _m.Called(bucketID, sourcePath, targetPath)
 
 	var r0 error
@@ -38,7 +41,7 @@ func (_m *Client) DownloadFile(bucketID string, sourcePath string, targetPath st
 }
 
 // ListFiles provides a mock function with given fields: bucketID
-func (_m *Client) ListFiles(bucketID string) ([]string, error) {
+func (_m *Client) ListFiles(ctx context.Context, bucketID string) ([]string, error) {
 	ret := _m.Called(bucketID)
 
 	var r0 []string
@@ -61,7 +64,7 @@ func (_m *Client) ListFiles(bucketID string) ([]string, error) {
 }
 
 // UploadFile provides a mock function with given fields: bucketID, sourcePath, targetPath
-func (_m *Client) UploadFile(bucketID string, sourcePath string, targetPath string) error {
+func (_m *Client) UploadFile(ctx context.Context, bucketID string, sourcePath string, targetPath string) error {
 	ret := _m.Called(bucketID, sourcePath, targetPath)
 
 	var r0 error
